changes to task 9

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,8 +7,8 @@
  */
 int main(void)
 {
-	unsigned int a = 2;
-	unsigned int n = 612852475143;
+	unsigned long int a = 2;
+	unsigned long int n = 612852475143;
 
 	while (a != n)
 	{
